requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-marshmallow
marshmallow-sqlalchemy
gunicorn
psycopg2-binary
requests
python-dotenv
